requirements.txt
Adafruit-Blinka==8.61.2
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-connectionmanager==3.1.5
adafruit-circuitpython-matrixkeypad==1.2.20
adafruit-circuitpython-requests==4.1.13
adafruit-circuitpython-typing==1.12.1
Adafruit-PlatformDetect==3.80.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
pyftdi==0.56.0
pyserial==3.5
pyusb==1.3.1
requests==2.32.4
RPi.GPIO==0.7.1
RPLCD==1.4.0
smbus2==0.5.0
sysv-ipc==1.1.0
typing_extensions==4.14.0
urllib3==2.5.0
